+++ idle-kluch-server/src/main/java/com/soze/idlekluch/kingdom/service/BuildingServiceImpl.java
@@ -235,7 +235,7 @@ public class BuildingServiceImpl implements BuildingService {
     //7. Check distance from source, if it's inside harvesting radius
 
     final float distance = EntityUtils.distance(harvester, source);
-    final float radius = harvesterComponent.getRadius();
+    final int radius = harvesterComponent.getRadius();
     if (distance > harvesterComponent.getRadius()) {
       throw new NoResourceInRadiusException(form.getMessageId(), "Distance: " + distance + ". Radius: " + radius);
     }
